+++ app/editor/Editor.jsx
@@ -1,5 +1,6 @@
 import PropTypes from 'prop-types';
 import React, { useCallback, useContext, useEffect } from 'react';
+import { Link } from 'react-router-dom';
 
 import { BangleEditor, useEditorState } from '@bangle.dev/react';
 import stopwatch from '@bangle.dev/react-stopwatch';
@@ -19,14 +20,8 @@ import {
   OrderedListButton,
   TodoListButton,
 } from '@bangle.dev/react-menu';
-import { config } from 'config/index';
-import {
-  specRegistry,
-  getPlugins,
-  menuKey,
-  emojiSuggestKey,
-} from 'editor/index';
-import { EditorManagerContext } from './EditorManager';
+import { menuKey, emojiSuggestKey } from 'editor/index';
+import { ExtensionEditorComponents } from 'bangle-io-context';
 
 const LOG = false;
 let log = LOG ? console.log.bind(console, 'play/Editor') : () => {};
@@ -36,9 +31,13 @@ Editor.propTypes = {
   editorId: PropTypes.number.isRequired,
 };
 
-export function Editor({ editorId, wsPath }) {
-  const { sendRequest, setEditor } = useContext(EditorManagerContext);
-
+export function Editor({
+  editorId,
+  wsPath,
+  bangleIOContext,
+  sendRequest,
+  setEditor,
+}) {
   useEffect(() => {
     log('mounting editor', editorId, wsPath);
     return () => {
@@ -47,8 +46,8 @@ export function Editor({ editorId, wsPath }) {
   }, [wsPath, editorId]);
 
   const plugins = useCallback(() => {
-    return getPlugins({ sendRequest, wsPath });
-  }, [sendRequest, wsPath]);
+    return bangleIOContext.getPlugins(wsPath, sendRequest);
+  }, [sendRequest, wsPath, bangleIOContext]);
 
   const onEditorReady = useCallback(
     (editor) => {
@@ -58,27 +57,20 @@ export function Editor({ editorId, wsPath }) {
   );
 
   const renderNodeViews = useCallback(
-    ({ node, updateAttrs, children, selected }) => {
-      // if (node.type.name === 'sticker') {
-      //   return (
-      //     <sticker.Sticker
-      //       node={node}
-      //       updateAttrs={updateAttrs}
-      //       selected={selected}
-      //     />
-      //   );
-      // }
-
+    (nodeViewRenderArg) => {
+      const { node, updateAttrs, children, selected } = nodeViewRenderArg;
       if (node.type.name === 'stopwatch') {
         return <stopwatch.Stopwatch node={node} updateAttrs={updateAttrs} />;
       }
+
+      return bangleIOContext.renderReactNodeViews(nodeViewRenderArg);
     },
-    [],
+    [bangleIOContext],
   );
 
   const editorState = useEditorState({
     plugins: plugins,
-    specRegistry,
+    specRegistry: bangleIOContext.specRegistry,
   });
 
   const renderMenuType = useCallback(({ type, menuKey }) => {
@@ -121,6 +113,7 @@ export function Editor({ editorId, wsPath }) {
     >
       <FloatingMenu menuKey={menuKey} renderMenuType={renderMenuType} />
       <EmojiSuggest emojiSuggestKey={emojiSuggestKey} />
+      <ExtensionEditorComponents bangleIOContext={bangleIOContext} />
     </BangleEditor>
   );
 }