+++ bangle-io-context/index.js
@@ -1,1 +1,2 @@
 export * from './bangle-io-context';
+export * from './ExtensionEditorComponents';