+++ extensions/collapsible-heading/index.js
@@ -1,0 +1,11 @@
+import { collapsibleHeadingDeco } from './collapsible-heading-deco';
+import { Extension } from 'extension-helpers';
+
+const extensionName = 'collapsible-heading';
+
+const extension = Extension.create({
+  name: extensionName,
+  editorPlugins: [collapsibleHeadingDeco.plugins()],
+});
+
+export default extension;