+++ js-libs/inline-palette/index.js
@@ -1,0 +1,4 @@
+import * as inlinePalette from './inline-palette';
+export * from './hooks';
+export { inlinePalette };
+export * from './inline-palette';